Don't import models and managers in __init__
This is deprecated as of django 1.8, and does not work as of django 1.9.
